requirements.txt
Django==4.2,<5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-ninja>=1.0
PyJWT>=2.8